requirements.txt
Django==2.2.6
django-crispy-forms==1.8.1
social-auth-app-django==3.1.0
social-auth-core==3.2.0
diff-match-patch==20200713
fontawesome-free==5.15.2
coolname==1.1.0
django-feather==0.2.6
celery==5.0.2
celery-progress==0.0.14
redis==3.5.3
mysqlclient==2.0.2
djangorestframework
pyyaml
requests
django-cors-headers